Commands: Fixed float part 2

+++ Artificial/TaskHandler.h
@@ -10,14 +10,14 @@ namespace Story
 	class TaskHandler final
 	{
 	public:
-		static const FloatMatrix& get_task_ins_by_name(const std::string& task)
+		static const DoubleMatrix& get_task_ins_by_name(const std::string& task)
 		{
 			if (!s_tasks.contains(task))
 				throw TaskNameException("(TaskHandler::get_task_ins_by_name) Task with name " + task + " does not exist");
 			return s_tasks[task]->get_ins();
 		}
 
-		static bool test_task_by_name(const std::string& task, const FloatMatrix& user_outs)
+		static bool test_task_by_name(const std::string& task, const DoubleMatrix& user_outs)
 		{
 			if (!s_tasks.contains(task))
 				throw TaskNameException("(TaskHandler::test_task_by_name) Task with name " + task + " does not exist");
